+++ config/sus.rb
@@ -2,3 +2,5 @@
 
 require "covered/sus"
 include Covered::Sus
+
+$LOAD_PATH.unshift(::File.expand_path("../fixtures", __dir__))